(UPDATE MINIMAL-v1.0) Fixed the font and fontSizing vars and added addLayer
FIX #1:
Fixed the canvas.font and canvas.fontSizing variables to work (removed the "var", not needed and needed to be removed for Jacks.js canvas.font and canvas.fontSizing to work)
FIX #2:
This part fixed a part of the canvas.text() function to also make canvas.font and canvas.fontSizing work.
FEATURE #1:
ADDED A Work-In-Progress function!:
canvas.addLayer();
This will simply create a new canvas with the parameters the user supplied.